+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - termux: use locally installed lsp servers
 - snippets: comment header & section snippets for lua
 - practicalli: toggle between Clojure src and test files
+- practicalli: showkeys visualise key presses in neovim
 
 ## Changed
 - lazy: enable zipPlugin to view src inside jar files